pyproject.toml
[tool.poetry]
name = "molecad"
version = "0.1.0"
description = "API for retrieving data from Pubchem databases and performing substructure search through local database"
authors = ["vvrubel <vasilyeva1510@gmail.com>"]
license = "free"

[tool.poetry.dependencies]
python = "^3.7.10"
requests = "^2.25.1"
types-requests = "^2.25.2"
pathlib = "^1.0.1"
mongordkit = { git = "git@github.com:rdkit/mongo-rdkit.git" }
rdkit-pypi = "^2021.3.3"
click = "^8.0.1"
click-pathlib = "^2020.3.13"
click-completion = "^0.5.2"
loguru = "^0.5.3"
importlib = "^1.0.4"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
fastapi = "^0.68.0"
uvicorn = "^0.14.0"
# TODO: add motor and other

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
mypy = "^0.910"

[tool.black]
line-length = 100
target_version = ["py37"]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
